requirements.txt
geocoder==1.38.1
geopy==2.3.0
matplotlib==3.6.2
numpy==1.23.5
opencv_python==4.6.0.66
pandas==1.5.2
Pillow==9.3.0
tensorflow==1.0.0
twilio==7.16.0
yagmail==0.15.293
